+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Additions
 * You can now smelt raw ore blocks in the Soul Blast Furnace
+* Added a config option to disable merging enchanted books
 
 ## 1.25.1
 ### Changes